+++ packages/parser/src/parsers/BigIntParser.ts
@@ -1,0 +1,15 @@
+import { BaseParser, ParseOptions } from "./Parser.js"
+
+export class BigIntParser extends BaseParser<bigint> {
+    override parseShape(value: unknown, options?: ParseOptions): bigint {
+        if (options?.coerce) {
+            const t = typeof value
+            const v = t === 'string' ? value as string : t === 'number' ? value as number : t === 'boolean' ? value as boolean : t === 'bigint' ? value as bigint : String(value)
+            value = BigInt(v)
+        }
+        if (typeof value !== 'bigint') {
+            throw new Error(`Expected bigint but got ${typeof value}`)
+        }
+        return value
+    }
+}
