+++ packages/parser/src/parsers/BooleanParser.ts
@@ -1,0 +1,13 @@
+import { BaseParser, ParseOptions } from "./Parser.js"
+
+export class BooleanParser extends BaseParser<boolean> {
+    override parseShape(value: unknown, options?: ParseOptions): boolean {
+        if (options?.coerce) {
+            value = Boolean(value)
+        }
+        if (typeof value !== 'boolean') {
+            throw new Error(`Expected boolean but got ${typeof value}`)
+        }
+        return value
+    }
+}
