+++ packages/parser/src/parsers/LazyParser.ts
@@ -1,0 +1,16 @@
+import { BaseParser, ParseOptions, Parser } from "./Parser.js"
+
+// Defers parsing until the last possible moment by wrapping a parser in a function. This allows for recursive data structure parsing.
+export class LazyParser<T> extends BaseParser<T> {
+    constructor(private parser: () => Parser<T>) {
+        super()
+    }
+
+    override parseShape(value: unknown, options?: ParseOptions): T {
+        return this.parser().parse(value, options)
+    }
+
+    override validate(value: T): void {
+        this.parser().validate(value)
+    }
+}
