+++ packages/parser/src/parsers/Parseable.ts
@@ -1,0 +1,6 @@
+import { Parser } from "./Parser.js";
+
+// Type that maps an object's properties to parsers for those properties
+export type Parseable<T> = {
+    [Property in keyof T]: Parser<T[Property]>
+}
