+++ packages/parser/src/parsers/UnknownParser.ts
@@ -1,0 +1,7 @@
+import { BaseParser, ParseOptions } from "./Parser.js";
+
+export class UnknownParser extends BaseParser<unknown> {
+    override parseShape(value: unknown, options?: ParseOptions): unknown {
+        return value
+    }
+}
