+++ packages/parser/src/tests/string.test.ts
@@ -1,0 +1,8 @@
+import { describe, expect, test, it } from 'vitest'
+
+
+describe("String", () => {
+    it("should parse a string", () => {
+        expect("hello").toEqual("hello");
+    });
+});
